Static Typing
We got int, float, string, and bool

implemented static typing.

created a typechecker class that runs on the AST before turning it into bytecode.

Declaration of variables is now done separately from assigning variables and error handling is there for when this is done incorrectly.

Isolated full pipeline into its own file that we can call using a reference to the .mars file we want to run.