Arrays and Matrices
This should be complete for the most part. But more testing should also be done.

Changed how assignment statements were being handled and added a new way of parsing the LHS for these assignment variables: (parse_assignable()).
(This handles the previous "." variables by flattening them into a single variable.
This handles a basic variable.
And it now additionally handles arrays)