Added Declaration for dictionaries to parser
Added a new function that better parses the type for when a variable or function is declared. parse_type()

This simplifies the parsing under
--- VAR & FUNC DECLARATION ---
and also allows for Dictionaries to be declared.

Still need to implement typechecking for dictionaries and now user defined classes/structs which the new parse_type() implements.

Still need to implement assigning Dictionary indices.

Still need to implement the bytecode and vm portion for dictionaries.