BUG INCLUDED! Dictionary Update (Declaring and More)
HUGE KNOWN BUG.
For some reason the VM is halted abruptly and too early and I do not know why.
Debug statements are currently placed in the VM file.
I believe the issue has todo with the GET_INDEX but I do not know why or how.

Other Fixes and Updates:

Still need to implement Dictionary Access and manipulation within the parser.

However I modified the bytecodegen so that it can now interpret DictLiterals.
It can also allow VarDecl for Dictionaries so we can define dictionaries.
I also reused and modified the ArrayAccess and Assign to also work with dictionaries.
The VM not interprets the index or key to decide whether it is an array or dictionary.
The VM now has renamed and adapted instructions called INDEX_GET and INDEX_SET, which are replacing ARRAY_GET and ARRAY_SET.

The typechecker has also been rewritten in order to handle dictionaries and a type checker and compatibility checker were added for better clarity and typechecking.

Also added 'COLON' to Lexer