Fixed Previous Bug with Arrays and Dicts
We can now do everything defined in the previous commit with no error.

Problem ended up being a break statement that was misused.

Added some error checking for premature breakage in the VM loop. And fixed a small bug for repeated imports (which was already caught in the typechecker but now also corrected in VM).

Still need to add assigning/modifying dicts.