Test Cases Implemented
Run All Test Files with:
python test_runner.py

Run A Specific Test File with:
python test_runner.py <test_filename.json>

1 file was created for the test_runner.py which runs all the tests

The a test folder was created in order to hold all of the test files which are written in json format for quick access and organization.

The interpreter file had a method added which runs the compiler based off a string instead of a filepath.

As seen there are already tests failing but many more tests should be created in order to cover everything, including error catching for bad code.

I added disabled testcases as well for testcases that I know don't work since we did not implement that logic in our compiler yet or because the test case isn't done being written or MOST IMPORTANTLY because the test case crashed the test_runner.

Please feel free to add any and as many test cases as you feel fit.